it builds

+++ src/util.h
@@ -138,7 +138,7 @@ static inline bjvm_utf8 hslc(heap_string str) {
 
 /// Converts the given null-terminated string to a slice. Use the STR macro for literals.
 static inline bjvm_utf8 str_to_utf8(const char *str) {
-  return (bjvm_utf8){.chars = (char *)str, .len = strlen(str)};
+  return (bjvm_utf8){.chars = (char *)str, .len = (int)strlen(str)};
 }
 
 #define fmt_slice(slice) (int)(slice).len, (slice).chars
